+++ sandbox/grommet-app/src/pages/layouts/kinds/OpsRamp/ResourceTraits.tsx
@@ -2,13 +2,11 @@ import React from 'react';
 import { Box, Text, Tag } from 'grommet';
 
 export const ResourceTraits = () => (
-  <>
-    <Box pad={{ vertical: 'small' }}>
-      <Text>Traits</Text>
-    </Box>
+  <Box gap="small">
+    <Text>Traits</Text>
     <Box direction="row" gap="xsmall">
       <Tag value="" name="Other" />
-      <Tag value="" name="Gateway is connected to this device" />
+      <Tag value="" name="Gateway is connected" />
     </Box>
-  </>
+  </Box>
 );
