+++ sandbox/grommet-app/src/pages/layouts/kinds/OpsRamp/index.ts
@@ -1,5 +1,6 @@
 export * from './navItems';
 export * from './OpsRamp';
+export * from './QuickFilters';
 export * from './ResourceAlerts';
 export * from './ResourceDetails';
 export * from './ResourceOverview';
